Light component rearranging in preparation for moving into subdirectories.
