+++ src/common/components/SwitchContent.jsx
