+++ src/common/index.js
@@ -1,9 +1,13 @@
 import * as utils from './utils';
+import Alert from './components/Alert';
 import PageLoading from './components/PageLoading';
+import SwitchContent from './components/SwitchContent';
 import { configureUserAccountApiService, fetchUserAccount } from './actions';
 
 export {
+  Alert,
   PageLoading,
+  SwitchContent,
   utils,
   configureUserAccountApiService,
   fetchUserAccount,
